Create HTTP Requests
Here is code for making HTTp requests to the  Cyber Security Open Source Intelligence websites. We need to add the functionality of using  the post method in websites so the user input (ip address) is entewred into the web site search query form there we can save the reuslts to a file and work on parsing the data.